Делает возможным добавить промо-фильм в избранное
